+++ CHANGES.md
@@ -3,6 +3,9 @@
 
 * allow the definition of `geographic_crs` used in the `geographic_bounds` property (https://github.com/cogeotiff/rio-tiler/pull/458)
 * use `contextlib.ExitStack` to better manager opening/closing rasterio dataset (https://github.com/cogeotiff/rio-tiler/pull/459)
+* moves `BBox, ColorTuple, Indexes, NoData, NumType` type definitions in `rio_tiler.types`
+* better types definition for ColorMap objects
+* fix some types issues
 
 # 3.0.0a4 (2021-11-10)